+++ test/tsrc/win_pct_test.c
@@ -17,8 +17,8 @@
 #include "test_constants.h"
 #include "test_util.h"
 
-void assert_win_pct_get(const float actual, const float expected) {
-  assert(within_epsilon(actual, expected));
+void assert_win_pct_get(const float actual, const double expected) {
+  assert(within_epsilon(actual, (float)expected));
 }
 
 void test_win_pct(void) {
